# chapter-6/final-lab/jobposting-back/react-openapi-generator/swagger.yaml
openapi: 3.0.1
info:
  title: Jobposting API
  description: Jobposting API documentation
  termsOfService: ''
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /api/job-postings/{jobId}:
    put:
      tags:
        - job-posting-resource
      operationId: updateJobPosting
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobPosting'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JobPosting'
    patch:
      tags:
        - job-posting-resource
      operationId: partialUpdateJobPosting
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobPosting'
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/JobPosting'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JobPosting'
...
  /api/job-postings:
    get:
      tags:
        - job-posting-resource
      operationId: getAllJobPostings
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            minimum: 0
            type: integer
            default: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            minimum: 1
            type: integer
            default: 20
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.'
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobPosting'
    post:
      tags:
        - job-posting-resource
      operationId: createJobPosting
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobPosting'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JobPosting'
...
  /api/job-postings/{id}:
    get:
      tags:
        - job-posting-resource
      operationId: getJobPosting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JobPosting'
    delete:
      tags:
        - job-posting-resource
      operationId: deleteJobPosting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
...
components:
  schemas:
...
    JobPosting:
      type: object
      properties:
        jobId:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        location:
          type: string
        postedDate:
          type: string
          format: date-time
...